move app and other file to container and components
