+++ src/containers/booklist.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import Book from './book';
+import Book from '../components/book';
 
 const BookList = ({ books }) => (
   <div>
